feat: drag-and-drop files on the ui (closes #74)
